Replacing std::vector with custom vector container

+++ tests/src/test_manager.hpp
@@ -2,9 +2,9 @@
 #include <vector>
 #include "defines.hpp"
 
-constexpr u8 TEST_FAIL = 0;
-constexpr u8 TEST_PASS = 1;
-constexpr u8 TEST_SKIP = 2;
+constexpr u8 TEST_FAIL = static_cast<u8>(0);
+constexpr u8 TEST_PASS = static_cast<u8>(1);
+constexpr u8 TEST_SKIP = static_cast<u8>(2);
 
 typedef u8 (*PFN_test)();
